% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCA_dimensionality_reduction.R
\name{DistSort}
\alias{DistSort}
\title{Sort OTUs/samples distance Matrix}
\usage{
DistSort(distance)
}
\arguments{
\item{distance}{distance matrix with samples at rows and OTUs at columns}
}
\value{
list of ranking of features by OTUs
}
\description{
Sort OTUs/samples distance Matrix
}
