% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCA_dimensionality_reduction.R
\name{pairDist}
\alias{pairDist}
\title{Importance Calculation}
\usage{
pairDist(x, y)
}
\arguments{
\item{x}{a matrix}

\item{y}{a matrix}
}
\value{
A Distance Matrix
}
\description{
Quickly calculate the importance in different OTUs
}
