% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCA_dimensionality_reduction.R
\name{run_mca_dim}
\alias{run_mca_dim}
\title{MicroMCA}
\usage{
run_mca_dim(X, nmcs = 10, reduction.name = "MCA")
}
\arguments{
\item{X}{Dataframe, Rows represent different OTUs or microorganisms classification entry and columns represent samples.}

\item{nmcs}{In order to estimate number of right singular vectors, In this case nmcs + 1 indicates number of right singular vectors.}

\item{reduction.name}{Method of the reduction default set to 'MCA' for feature data and mca.}
}
\description{
Dimensionality reduction of high-dimensional microbiome data using MCA
}
\details{
The selected parameters are adapted from Griffiths & Steyvers (2004).
}
\examples{

}
